extract validation logic from filter class
