+++ lib/filterable.rb
@@ -3,6 +3,7 @@ require 'filterable/filter_validator'
 require 'filterable/filtrator'
 require 'filterable/sort'
 require 'filterable/subset_comparator'
+require 'filterable/type_validator'
 
 module Filterable
   extend ActiveSupport::Concern
